+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 -   Update minimum Java to Java 8
 -   Add in Discord rich presence integration
+-   Fix post launch error check dialog not being displayed correctly